+++ extensions/messages/src/components/MessageActions.tsx
@@ -2,7 +2,7 @@ import { Action, ActionPanel, AI, Icon, Keyboard, open, showToast, Toast, useNav
 import { showFailureToast, useLocalStorage } from "@raycast/utils";
 
 import { extractOTP, getAttachmentType, getMessagesUrl } from "../helpers";
-import { Message } from "../hooks/useMessages";
+import type { Message } from "../types";
 
 import Details from "./Details";
 import Instructions from "./Instructions";
